+++ docs/changelog.rst
@@ -3,6 +3,9 @@ Changelog
 Unreleased
 ^^^^^^^^^^
 - Updated DFXPReader to ignore paragraphs that only contain spaces, tabs or new lines
+- Added CaptionReadTimingError for invalid SCC timestamps
+- Added CaptionReadSyntaxError for invalid colors in SAMIReader
+- Raise CaptionReadTimingError when missing 'begin' or 'end' and 'dur' time on DFXPReader
 
 2.0.4
 ^^^^^